Bugfix: crash when two modals follow each other
After introducing modal animation (for mobile), I introduced a race condition between the dismiss animation and the promise return of the first modal. As a result, the then() actions were starting to trigger the second modal before the first modal was dismissed.
Solve: I made the hide animation a promise that resolves after the modal is dismissed, and I make the modal then() action trigger only after the hide() promise returns.