Fixed: rentering of the minigraph with wrong line color
This happens because on Minigraph.js, the mid point of the gradient is calculating based on the line that includes the projection while the victory line renders the gradient on the graph without the projection. The only way the graph scales properly is to represent the full data set (including the projection) but there is a limitation of victory lines that you can't style individual segments since the line is styled once. To fix it, we're scaling the gradient based on the line without the projection.