Fix: on mobile, tapping on the future placement of the + or edit button triggers the action.
on mobile, the tap-as-hover event is indistinguishable from the tap event so 2 events get fired when tapping directly on one of the tools button... resulting in entering edit mode or factory mode even when the buttons were not visible. By slightly delaying the appearance of the tool buttons, we're able to suppress this effect because the extra tap event happens while the buttons aren't visible yet, and therefore doesn't trigger anything.