Bank UI: Decouple aggregator steps
To avoid bloating the aggregator connection component for each aggregator implemented, we can now do it via inheritance of the AggregatorConnectorStep class and override the default behavior.
To achieve this, the state machine can now specify a renderable as a function taking the Context as argument instead of a direct component. Both are accepted.